Updated the script file so it runs the project tool before hand

+++ projects/upgrade_project.sh
@@ -4,14 +4,21 @@ set -ex
 if [ -z $PROJECTTOOL ]; then
 	PROJECTTOOL=/e/Source/ProjectTool/Source/bin/Debug/net8.0/ProjectTool
 fi
-BASE=`pwd`
 
-$PROJECTTOOL SHOWVERSIONEDTYPES DESTINATION=resourceslist.json
+if [ -z $PROJECTTOOL ]; then
+	echo "Environment variable CORERESOURCES_DLL is not set skipping convertion!"
+	exit 0
+fi
+
+BASE=`pwd`
+mkdir Prefabs
+
+$PROJECTTOOL SHOWVERSIONEDTYPES DESTINATION=resourceslist.json SOURCE="$CORERESOURCES_DLL"
 
 for project in *; do
 	if [ -d $project ]; then
 		cd $project
-		$PROJECTTOOL project save format=versioned source=./$project.yyp forcevers0=true RESOURCETYPESPATH=$BASE/resourceslist.json PREFABSFOLDER=.
+		$PROJECTTOOL project save format=versioned source=./$project.yyp RESOURCETYPESPATH=$BASE/resourceslist.json PREFABSFOLDER=$BASE/Prefabs
 		rm $project.resource_order
 		cd ..
 	fi
@@ -19,3 +26,4 @@ done
 
 cd $BASE
 rm resourceslist.json
+rm -rf Prefabs
